refactor: Renamed and Reorganized some files for improved organization, and readability.
