+++ app/hotels/[hotelId]/components/overview/FeaturedFacilities.jsx
