+++ app/hotels/[hotelId]/components/overview/FeaturedPolicies.jsx
