+++ app/hotels/[hotelId]/components/overview/NearbyInterests.jsx
